+++ src/utils/string.ts
@@ -47,3 +47,13 @@ export const slugifiedCollectionIdToCollectionId = (slugifiedCollectionId: strin
   }
   return splits[splits.length - 1];
 };
+
+/**
+ * Cleans a transcript by removing left-to-right and right-to-left marks
+ *
+ * @param {string} text - The text to clean
+ * @returns {string} The cleaned text
+ */
+export const cleanTranscript = (text: string): string => {
+  return text.replace(/[\u200E\u200F]/g, '');
+};
